Basic demos now work
