+++ demo/circles/index.html
@@ -10,7 +10,7 @@
 <circles-demo></circles-demo>
 <script src="../../window.js"></script>
 <script>
-  fritz.app = new Worker('./app.js');
+  fritz.use(new Worker('./app.js'));
 </script>
 </body>
 </html>
