+++ src/message-types.js
@@ -1,0 +1,4 @@
+export const DEFINE = 'define';
+export const TRIGGER = 'trigger';
+export const RENDER = 'render';
+export const EVENT = 'event';
