--- src/window/trigger.js
@@ -1,9 +1,0 @@
-import { getInstance } from './util.js';
-
-export default function(msg, fritz) {
-  let inst = getInstance(msg.id, fritz);
-  let event = new Event(msg.event.type, {
-    bubbles: true
-  });
-  inst.dispatchEvent(event);  
-};
