--- src/window/util.js
@@ -1,3 +1,0 @@
-export function getInstance(id, fritz){
-  return fritz._instances[id];
-};
