--- src/worker/is-promise.js
@@ -1,5 +1,0 @@
-function isPromise(obj) {
-  return !!obj && (typeof obj === 'object' || typeof obj === 'function') && typeof obj.then === 'function';
-}
-
-export default isPromise;
