--- src/worker/util.js
@@ -1,3 +1,0 @@
-export function getInstance(fritz, id){
-  return fritz._instances[id];
-};
