+++ test/events/index.html
@@ -1,6 +1,6 @@
 <!doctype html>
 <script src="../../window.js"></script>
 <script>
-  fritz.app = new Worker('./app.js');
+  fritz.use(new Worker('./app.js'));
 </script>
 <event-element></event-element>
